Fixing types to please -Werror

+++ src/include/SRProxy/BasicTypesProxy.h
@@ -176,7 +176,7 @@ protected:
   std::string IndexField() const;
 
   /// add [i], or something more complex for nested CAFs
-  std::string Subscript(int i) const;
+  std::string Subscript(size_t i) const;
 
   std::string SubName() const;
 
@@ -401,7 +401,7 @@ public:
   }
 
 protected:
-  void EnsureElem(int i) const {
+  void EnsureElem(size_t i) const {
     CheckIndex(i, N);
     if (fElems[i]) {
       return; // element already created
